requirements.txt
beautifulsoup4~=4.9.1
PyYAML~=5.3.1
six~=1.15.0
flask~=1.1.2
requests~=2.23.0
rsa~=4.0
coverage~=5.1
jsonschema~=3.2.0
aiohttp~=3.6.2
dataclasses~=0.6
cryptography~=2.9.2
psutil~=5.7.0
python-dotenv~=0.13.0
netifaces~=0.10.9
apscheduler~=3.6.3
passlib~=1.7.2
jinja2~=2.11.2
SQLAlchemy~=1.3.17
Werkzeug~=1.0.1
jinja2schema~=0.1.4
RestrictedPython~=5.0
setuptools~=47.3.1
click~=7.1.2
python-dateutil~=2.8.1
tzlocal~=2.1
Pygments~=2.6.1
pyfakefs~=4.0.2
responses~=0.10.14
aioresponses~=0.6.4
asynctest~=0.13.0
prompt_toolkit
docopt
gunicorn